edit auth

+++ app/Http/Controllers/Api/ACL/FriendController.php
@@ -8,6 +8,7 @@ use App\Repositories\ACL\LogRepository;
 use App\Repositories\ACL\UserRepository;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
+use Illuminate\Support\Facades\Auth;
 
 
 class FriendController extends Controller
@@ -32,13 +33,24 @@ class FriendController extends Controller
         if (!$user_receive) {
             return response(['status' => 0, 'data' => array(), 'message' => 'رقم المستخدم المرسل اليه خطاء'], 400);
         }
-        $friend = new Friend();
-        $friend->user_send_id = $user_send->id;
-        $friend->user_receive_id = $user_receive->id;
-        $friend->status = 0;
-        $friend->save();
-        $this->logRepository->Create_Data('' . $user_send->id . '', 'ارسال', 'ارسال طلب صداقه');
-        return response(['status' => 1, 'data' => ['request_friend_number' => $friend->id], 'message' => 'تم ارسال طلب الصداقه'], 200);
+        if ($user_send->id == Auth::User()->id) {
+            if ($user_send->status == 0) {
+                return response(['status' => 0, 'data' => array(), 'message' => 'برجاء الاتصال بخدمه العملاء'], 400);
+            }
+            $friend = Friend::where('user_send_id', $user_send->id)->where('user_receive_id', $user_receive->id)
+                ->orwhere('user_receive_id', $user_send->id)->where('user_send_id', $user_receive->id)->first();
+            if (!$friend) {
+                $friend = new Friend();
+                $friend->user_send_id = $user_send->id;
+                $friend->user_receive_id = $user_receive->id;
+                $friend->status = 0;
+                $friend->save();
+                $this->logRepository->Create_Data('' . $user_send->id . '', 'ارسال', 'ارسال طلب صداقه');
+                return response(['status' => 1, 'data' => ['request_friend_number' => $friend->id], 'message' => 'تم ارسال طلب الصداقه'], 200);
+            }
+            return response(['status' => 0, 'data' => array(), 'message' => 'يوجد صداقه مسبقا'], 400);
+        }
+        return response(['status' => 0, 'data' => array(), 'message' => 'لا يمكن اتمام الطلب'], 400);
     }
 
     public function accept_friend(Request $request)
@@ -49,10 +61,13 @@ class FriendController extends Controller
         }
         $friend = Friend::find($request->request_friend_number);
         if ($friend) {
-            $friend->status = 1;
-            $friend->update();
-            $this->logRepository->Create_Data('' . $user->id . '', 'قبول', 'قبول طلب صداقه');
-            return response(['status' => 1, 'data' => array(), 'message' => 'تم قبول طلب صداقه'], 200);
+            if ($user->id == Auth::user()->id && $friend->user_receive_id == Auth::user()->id) {
+                $friend->status = 1;
+                $friend->update();
+                $this->logRepository->Create_Data('' . $user->id . '', 'قبول', 'قبول طلب صداقه');
+                return response(['status' => 1, 'data' => array(), 'message' => 'تم قبول طلب صداقه'], 200);
+            }
+            return response(['status' => 0, 'data' => array(), 'message' => 'لا يمكن اتمام الطلب'], 400);
         }
         return response(['status' => 0, 'data' => array(), 'message' => 'خطا فى تحميل البيانات طلب الصداقه'], 400);
 
@@ -64,11 +79,17 @@ class FriendController extends Controller
         if (!$user) {
             return response(['status' => 0, 'data' => array(), 'message' => 'خطا فى تحميل البيانات المستخدم'], 400);
         }
+        if ($user->status == 0) {
+            return response(['status' => 0, 'data' => array(), 'message' => 'برجاء الاتصال بخدمه العملاء'], 400);
+        }
         $friend = Friend::find($request->request_friend_number);
         if ($friend) {
-            $friend->delete();
-            $this->logRepository->Create_Data('' . $user->id . '', 'مسح', 'مسح  الصداقه');
-            return response(['status' => 1, 'data' => array(), 'message' => 'تم الغاء الصداقه'], 200);
+            if ($user->id == Auth::user()->id && ($friend->user_send_id == Auth::user()->id || $friend->user_receive_id == Auth::user()->id)) {
+                $friend->delete();
+                $this->logRepository->Create_Data('' . $user->id . '', 'مسح', 'مسح  الصداقه');
+                return response(['status' => 1, 'data' => array(), 'message' => 'تم الغاء الصداقه'], 200);
+            }
+            return response(['status' => 0, 'data' => array(), 'message' => 'لا يمكن اتمام الطلب'], 400);
         }
         return response(['status' => 0, 'data' => array(), 'message' => 'خطا فى تحميل البيانات'], 400);
 
@@ -79,6 +100,9 @@ class FriendController extends Controller
         $user = $this->userRepository->Get_One_Data($request->user_id);
         if (!$user) {
             return response(['status' => 0, 'data' => array(), 'message' => 'خطا فى تحميل البيانات المستخدم'], 400);
+        }
+        if ($user->status == 0) {
+            return response(['status' => 0, 'data' => array(), 'message' => 'برجاء الاتصال بخدمه العملاء'], 400);
         }
         $friend = Friend::where('user_send_id', $user->id)->where('status', 1)
             ->orwhere('user_receive_id', $user->id)->where('status', 1)
@@ -100,6 +124,9 @@ class FriendController extends Controller
         $user = $this->userRepository->Get_One_Data($request->user_id);
         if (!$user) {
             return response(['status' => 0, 'data' => array(), 'message' => 'خطا فى تحميل البيانات المستخدم'], 400);
+        }
+        if ($user->status == 0) {
+            return response(['status' => 0, 'data' => array(), 'message' => 'برجاء الاتصال بخدمه العملاء'], 400);
         }
         $request_friend = Friend::where('user_receive_id', $user->id)->where('status', 0)->get();
         if ($request->status_auth == 1) {
